factor out clock, improve selection ui
